+++ src/Components/RegisterBarbershop/RBS.scss
@@ -4,7 +4,6 @@
     padding: 2vh 4vw;
     align-items: center;
 
-
     h1,
     h4 {
         padding: 1vh 0;
