+++ src/Components/Stepper/StepperHeader.tsx
@@ -1,20 +1,18 @@
 import React, { Component } from "react";
 import { Step, StepLabel, Stepper } from "@material-ui/core";
-import './styles.scss';
-
+import "./styles.scss";
 
 interface StepperProps {
     currentStep: number;
-    stepLabels: string[]
+    stepLabels: string[];
 }
 
-class StepperHeader extends Component<StepperProps,{}> {
-
+class StepperHeader extends Component<StepperProps, {}> {
     render() {
         return (
             <div className="stepper-container">
                 <Stepper activeStep={this.props.currentStep} alternativeLabel>
-                    {this.props.stepLabels.map((label =>
+                    {this.props.stepLabels.map((label) => (
                         <Step key={label}>
                             <StepLabel>{label}</StepLabel>
                         </Step>
@@ -24,6 +22,5 @@ class StepperHeader extends Component<StepperProps,{}> {
         );
     }
 }
-
 
 export default StepperHeader;
