+++ src/Components/Stepper/styles.scss
@@ -1,3 +1,3 @@
 .MuiStepper-horizontal {
-  background-color: lightgray  !important;
+    background-color: lightgray !important;
 }
