+++ src/pages/RBSMain.tsx
@@ -1,22 +1,24 @@
 import React, { Component } from "react";
-import StepperHeader from "../Stepper/StepperHeader";
-import { ADD_BARBERSHOP_STEPS } from "./constants";
+import StepperHeader from "../components/Stepper/StepperHeader";
+import { ADD_BARBERSHOP_STEPS } from "../types/constants";
 import { Switch, Route } from "react-router-dom";
-import RBSFormRegister from "./RBSFormRegister";
+import RBSFormRegister from "../components/RegisterBarbershop/RBSFormRegister";
 
 interface RBSMainProps {
     currentStep: number;
 }
 
 class RBSMain extends Component<RBSMainProps, {}> {
-
     render() {
         return (
             <div>
-                <StepperHeader currentStep={this.props.currentStep} stepLabels={ADD_BARBERSHOP_STEPS}/>
+                <StepperHeader
+                    currentStep={this.props.currentStep}
+                    stepLabels={ADD_BARBERSHOP_STEPS}
+                />
                 <Switch>
                     <Route path="/createshop/register">
-                        <RBSFormRegister/>
+                        <RBSFormRegister />
                     </Route>
                     <Route path="/createshop/hours"></Route>
                     <Route path="/createshop/confirm"></Route>
