--- src/reducers/RBFormReducer.ts
@@ -1,3 +1,0 @@
-// TODO: add register barbershop form reducer
-
-export const todo = "todo";
